Docker

+++ server/database.sql
@@ -1,7 +1,7 @@
-CREATE TABLE items (
+CREATE TABLE items(
   id SERIAL PRIMARY KEY,
   name TEXT,
   description TEXT,
-  price INTEGER,
-  quantity INTEGER,
+  price FLOAT,
+  quantity INTEGER
 );
